# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malaika-booking API",
        "description": "",
        "contact": {
            "email": "kirian@fruntera.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Lista de viajes",
                "description": "Lista los viajes",
                "operationId": "tripList",
                "parameters": [
                    {
                        "name": "trip_state",
                        "in": "query",
                        "description": "Id del estado del viaje",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "client",
                        "in": "query",
                        "description": "Id de cliente",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TripListResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Retorna un viaje por ID",
                "description": "Retorna un viaje",
                "operationId": "getTripById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de viaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TripResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trips"
                ],
                "summary": "Actualiza los datos del viaje",
                "description": "Actualiza los datos del viaje",
                "operationId": "updateTrip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de viaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update trip",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "warehouse_id",
                                    "stock"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tour por Italia"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tour por Italia"
                                    },
                                    "commentary": {
                                        "type": "string",
                                        "example": "Es un viaje muy chulo, comes pasta y pizza"
                                    },
                                    "trip_state_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trip updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TripListResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Trip title",
                        "type": "string",
                        "example": "Antartida desde el mar"
                    },
                    "description": {
                        "description": "Trip description",
                        "type": "string",
                        "example": "Es un viaje muy chulo"
                    },
                    "commentary": {
                        "description": "Trip commentary",
                        "type": "string",
                        "example": "Especial para ciegos"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TripResource"
                }
            },
            "TripResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Trip title",
                        "type": "string",
                        "example": "Antartida desde el mar"
                    },
                    "description": {
                        "description": "Trip description",
                        "type": "string",
                        "example": "Es un viaje muy chulo"
                    },
                    "commentary": {
                        "description": "Trip commentary",
                        "type": "string",
                        "example": "Especial para ciegos"
                    },
                    "state": {
                        "description": "Trip state",
                        "type": "string",
                        "example": "OPEN"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TripResource"
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}